src/test/java/br/com/dbc/vemser/projetoTelecomunicacoes/service/FaturaServiceTest.java
package br.com.dbc.vemser.projetoTelecomunicacoes.service;

import br.com.dbc.vemser.projetoTelecomunicacoes.dto.FaturaDTO;
import br.com.dbc.vemser.projetoTelecomunicacoes.dto.PagamentoDTO;
import br.com.dbc.vemser.projetoTelecomunicacoes.entity.Fatura;
import br.com.dbc.vemser.projetoTelecomunicacoes.mocks.ClienteMock;
import br.com.dbc.vemser.projetoTelecomunicacoes.mocks.FaturaMock;
import br.com.dbc.vemser.projetoTelecomunicacoes.mocks.PagamentoMock;
import br.com.dbc.vemser.projetoTelecomunicacoes.repository.FaturaRepository;
import com.fasterxml.jackson.databind.DeserializationFeature;
import com.fasterxml.jackson.databind.ObjectMapper;
import com.fasterxml.jackson.databind.SerializationFeature;
import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.test.util.ReflectionTestUtils;

import java.util.Optional;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.ArgumentMatchers.anyInt;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
class FaturaServiceTest {


    @InjectMocks
    private FaturaService faturaService;
    @Mock
    private ObjectMapper objectMapper;
    @Mock
    private FaturaRepository faturaRepository;

    private FaturaMock faturaMock;

    private ClienteMock clienteMock;

    private PagamentoMock pagamentoMock;

    @BeforeEach
    void init(){
        faturaMock = new FaturaMock();
        clienteMock = new ClienteMock();
        ReflectionTestUtils.setField(faturaService, "objectMapper", getObjectMapperInstance());
    }

    @Test
    void devePagarFaturaComSucesso() throws Exception {
        Integer id = 1;
        pagamentoMock = new PagamentoMock();
        PagamentoDTO pagamentoDTO = pagamentoMock.retornaPagamentoDTO(1);
        Fatura fatura = faturaMock.retornarEntidadeFatura(1);

        when(faturaRepository.findById(anyInt())).thenReturn(Optional.of(fatura));
        when(faturaRepository.save(any(Fatura.class))).thenReturn(fatura);

        FaturaDTO faturaPaga = faturaService.pagarFatura(id, pagamentoDTO);

        assertNotNull(faturaPaga);
        assertEquals(faturaPaga.getDataBaixa(), pagamentoDTO.getDataBaixa());
        assertEquals(faturaPaga.getValorPago(), pagamentoDTO.getValorPago());



    }


    public static ObjectMapper getObjectMapperInstance() {

        ObjectMapper objectMapper = new ObjectMapper();

        objectMapper.registerModule(new JavaTimeModule());
        objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
        objectMapper.disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);

        return objectMapper;
    }

}